requirements.txt
pip
selenium==3.141.0
schedule==0.6.0
Pillow==7.0.0